remove any mention of old status bars also moved FancyStatusBars.java
