+++ src/main/java/de/hysky/skyblocker/mixin/WindowMixin.java
@@ -1,6 +1,6 @@
 package de.hysky.skyblocker.mixin;
 
-import de.hysky.skyblocker.skyblock.FancyStatusBars;
+import de.hysky.skyblocker.skyblock.fancybars.FancyStatusBars;
 import net.minecraft.client.util.Window;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
