+++ src/main/resources/assets/skyblocker/lang/en_us.json
@@ -26,14 +26,6 @@
   "text.autoconfig.skyblocker.option.general.bars": "Health, Mana, Defence & XP Bars",
   "text.autoconfig.skyblocker.option.general.bars.enableBars": "Enable Bars",
   "text.autoconfig.skyblocker.option.general.bars.barpositions": "Configure Bar Positions",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.LAYER1": "Layer 1",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.LAYER2": "Layer 2",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.RIGHT": "Right",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.NONE": "Disabled",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.healthBarPosition": "Health Bar Position",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.manaBarPosition": "Mana Bar Position",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.defenceBarPosition": "Defence Bar Position",
-  "text.autoconfig.skyblocker.option.general.bars.barpositions.experienceBarPosition": "Experience Bar Position",
   "text.autoconfig.skyblocker.option.general.experiments": "Experiments Solver",
   "text.autoconfig.skyblocker.option.general.experiments.enableChronomatronSolver": "Enable Chronomatron Solver",
   "text.autoconfig.skyblocker.option.general.experiments.enableSuperpairsSolver": "Enable Superpairs Solver",
